[유재우] feature/entity/fix: 최종 파일구조 및 필드명, 엔티티 수정 완
