+++ src/main/java/com/hugudungs/hugupjigup/common/enums/CommentType.java
@@ -1,4 +1,4 @@
-package com.hugudungs.hugupjigup.data.entity.enums;
+package com.hugudungs.hugupjigup.common.enums;
 
 public enum CommentType {
  COMMENT,
