+++ src/main/java/com/hugudungs/hugupjigup/common/enums/Gender.java
@@ -1,0 +1,6 @@
+package com.hugudungs.hugupjigup.common.enums;
+
+public enum Gender {
+    MALE,
+    FEMALE
+}
