+++ src/main/java/com/hugudungs/hugupjigup/data/entity/board/BoardTypeEntity.java
@@ -1,22 +1,19 @@
 package com.hugudungs.hugupjigup.data.entity.board;
 
 import com.hugudungs.hugupjigup.data.entity.common.BaseEntity;
-import com.hugudungs.hugupjigup.data.entity.enums.BoardType;
+import com.hugudungs.hugupjigup.common.enums.BoardType;
+
+import jakarta.persistence.EnumType;
 import jakarta.persistence.AttributeOverride;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
 import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.NoArgsConstructor;
 
 @Entity
 @Table(name = "board_types")
-@NoArgsConstructor
-@AllArgsConstructor
 @AttributeOverride(name = "id", column = @Column(name = "board_type_id", columnDefinition = "TINYINT"))
-public class BoardTypes extends BaseEntity {
+public class BoardTypeEntity extends BaseEntity {
     @Enumerated(EnumType.STRING)
     @Column(name = "board_type")
     private BoardType boardType;
