+++ src/main/java/com/hugudungs/hugupjigup/data/entity/board/Free.java
@@ -1,14 +1,18 @@
 package com.hugudungs.hugupjigup.data.entity.board;
 
 import com.hugudungs.hugupjigup.data.entity.user.User;
-import jakarta.persistence.*;
-import lombok.*;
-import java.util.Date;
+
+import jakarta.persistence.FetchType;
+import jakarta.persistence.AttributeOverride;
+import jakarta.persistence.AttributeOverrides;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
 
 @Entity
 @Table(name = "free")
-@Getter
-@Setter
 @AttributeOverrides({
         @AttributeOverride(name = "id", column = @Column(name = "free_id", nullable = false)),
         @AttributeOverride(name = "title", column = @Column(name = "free_title", nullable = false)),
@@ -16,11 +20,11 @@ import java.util.Date;
         @AttributeOverride(name = "views", column = @Column(name = "free_views", nullable = false, columnDefinition = "INT DEFAULT 0"))
 })
 public class Free extends BaseBoardEntity{
-    @ManyToOne
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "board_type_id", nullable = false)
-    private BoardTypes boardTypeId;
+    private BoardTypeEntity boardType;
 
-    @ManyToOne
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id", nullable = false)
-    private User userId;
+    private User user;
 }
