+++ src/main/java/com/hugudungs/hugupjigup/data/entity/comment/FreeComment.java
@@ -2,10 +2,17 @@ package com.hugudungs.hugupjigup.data.entity.comment;
 
 import com.hugudungs.hugupjigup.data.entity.board.Free;
 import com.hugudungs.hugupjigup.data.entity.user.User;
-import jakarta.persistence.*;
 
-import java.util.Date;
-import lombok.*;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.AttributeOverride;
+import jakarta.persistence.AttributeOverrides;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import lombok.Getter;
+import lombok.Setter;
 
 @Entity
 @Table(name = "free_comment")
@@ -16,16 +23,15 @@ import lombok.*;
         @AttributeOverride(name = "content", column = @Column(name = "free_comment_content", nullable = false, columnDefinition = "TEXT"))
 })
 public class FreeComment extends BaseCommentEntity {
-
-    @ManyToOne
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "free_id", nullable = false)
     private Free free;
 
-    @ManyToOne
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id", nullable = false)
-    private User userId;
+    private User user;
 
-    @ManyToOne
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "comment_type_id", nullable = false)
     private CommentTypes commentTypeId;
 }
