--- src/main/java/com/hugudungs/hugupjigup/data/entity/enums/Gender.java
@@ -1,6 +1,0 @@
-package com.hugudungs.hugupjigup.data.entity.enums;
-
-public enum Gender {
-    MALE,
-    FEMALE
-}
