--- src/main/java/com/hugudungs/hugupjigup/data/entity/enums/LoginType.java
@@ -1,6 +1,0 @@
-package com.hugudungs.hugupjigup.data.entity.enums;
-
-public enum LoginType {
-    COMMON,
-    SOCIAL
-}
