--- src/main/java/com/hugudungs/hugupjigup/data/entity/enums/ProfileType.java
@@ -1,6 +1,0 @@
-package com.hugudungs.hugupjigup.data.entity.enums;
-
-public enum ProfileType {
-    MENTOR,
-    MENTEE
-}
