+++ src/main/java/com/hugudungs/hugupjigup/data/entity/matching/Matching.java
@@ -1,11 +1,13 @@
-package com.hugudungs.hugupjigup.data.entity.board;
+package com.hugudungs.hugupjigup.data.entity.matching;
 
+import com.hugudungs.hugupjigup.data.entity.board.BaseBoardEntity;
 import com.hugudungs.hugupjigup.data.entity.user.User;
+
+import jakarta.persistence.FetchType;
 import jakarta.persistence.AttributeOverride;
 import jakarta.persistence.AttributeOverrides;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
@@ -22,9 +24,6 @@ public class Matching extends BaseBoardEntity {
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id", referencedColumnName = "user_id")
     private User user;
-
-    @Column(name = "user_id", insertable = false, updatable = false)
-    private Long userId;
 
     @Column(nullable = false)
     private String tag; // 태그 (예: 작업, 기술)
