+++ src/main/java/com/hugudungs/hugupjigup/data/entity/user/User.java
@@ -1,13 +1,14 @@
 package com.hugudungs.hugupjigup.data.entity.user;
 
 import com.hugudungs.hugupjigup.data.entity.common.BaseEntity;
-import com.hugudungs.hugupjigup.data.entity.enums.LoginType;
+import com.hugudungs.hugupjigup.common.enums.LoginType;
+
+import jakarta.persistence.EnumType;
+import jakarta.persistence.FetchType;
 import jakarta.persistence.AttributeOverride;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
-import jakarta.persistence.FetchType;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
@@ -18,10 +19,9 @@ import java.time.LocalDateTime;
 @Table(name = "users")
 @AttributeOverride(name = "id", column = @Column(name = "user_id"))
 public class User extends BaseEntity {
-
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "role_type_id", referencedColumnName = "role_type_id")
-    private RoleTypes roleTypesId;
+    private RoleTypeEntity roleType;
 
     @Column(name = "user_nickname", nullable = false, unique = true)
     private String nickName;
